feat(kuma-cp) create separate package for tags
