+++ pkg/xds/envoy/listeners/util_test.go
@@ -2,8 +2,6 @@ package listeners_test
 
 import (
 	"errors"
-	"github.com/Kong/kuma/api/mesh/v1alpha1"
-	"regexp"
 
 	envoy_type "github.com/envoyproxy/go-control-plane/envoy/type"
 	"github.com/golang/protobuf/ptypes/wrappers"
@@ -202,7 +200,7 @@ var _ = Describe("ConvertPercentage", func() {
 		input    *wrappers.DoubleValue
 		expected *envoy_type.FractionalPercent
 	}
-	DescribeTable("shoudl properly converts from percent to fractional percen",
+	DescribeTable("should properly converts from percent to fractional percen",
 		func(given testCase) {
 			fpercent := ConvertPercentage(given.input)
 			Expect(fpercent).To(Equal(given.expected))
@@ -256,50 +254,4 @@ var _ = Describe("ConvertBandwidth", func() {
 			expected: 120000000,
 		}),
 	)
-})
-
-var _ = Describe("ConvertTags", func() {
-	type testCase struct {
-		serviceTags v1alpha1.MultiValueTagSet
-		matchTags   v1alpha1.SingleValueTagSet
-		expected    bool
-	}
-	DescribeTable("should generate regex for matching service's tags",
-		func(given testCase) {
-			// when
-			regexStr := ConvertTags(given.matchTags)
-			re, err := regexp.Compile(regexStr)
-			// then
-			Expect(err).ToNot(HaveOccurred())
-			// when
-			matched := re.MatchString(" " + given.serviceTags.String())
-			// then
-			Expect(matched).To(Equal(given.expected))
-		},
-		Entry("match without middle tag2", testCase{
-			serviceTags: v1alpha1.MultiValueTagSet{
-				"tag1": {"value1": true, "value2": true},
-				"tag2": {"value2": true, "value3": true},
-				"tag3": {"value3": true, "value4": true},
-			},
-			matchTags: v1alpha1.SingleValueTagSet{
-				"tag1": "value1",
-				"tag3": "value3",
-			},
-			expected: true,
-		}),
-		Entry("doesn't match", testCase{
-			serviceTags: v1alpha1.MultiValueTagSet{
-				"tag1": {"value1": true, "value2": true},
-				"tag2": {"value2": true, "value3": true},
-				"tag3": {"value3": true, "value4": true},
-			},
-			matchTags: v1alpha1.SingleValueTagSet{
-				"tag1": "value1",
-				"tag3": "value5",
-			},
-			expected: false,
-		}),
-	)
-
 })
