+++ pkg/xds/envoy/routes/tags_configurer.go
@@ -1,11 +1,10 @@
 package routes
 
 import (
-	"fmt"
-	"strings"
-
 	envoy_api_v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
+
+	"github.com/Kong/kuma/pkg/xds/envoy/tags"
 
 	mesh_proto "github.com/Kong/kuma/api/mesh/v1alpha1"
 )
@@ -29,18 +28,7 @@ func (t *TagsHeaderConfigurer) Configure(rc *envoy_api_v2.RouteConfiguration) er
 		return nil
 	}
 	rc.RequestHeadersToAdd = append(rc.RequestHeadersToAdd, &envoy_core.HeaderValueOption{
-		Header: &envoy_core.HeaderValue{Key: TagsHeaderName, Value: serializeTags(t.tags)},
+		Header: &envoy_core.HeaderValue{Key: TagsHeaderName, Value: tags.Serialize(t.tags)},
 	})
 	return nil
-}
-
-func serializeTags(tags mesh_proto.MultiValueTagSet) string {
-	var pairs []string
-	for _, key := range tags.Keys() {
-		pairs = append(pairs, fmt.Sprintf("%s=%s", key, strings.Join(tags.Values(key), ",")))
-	}
-	if len(pairs) == 0 {
-		return ""
-	}
-	return "&" + strings.Join(pairs, "&") + "&"
 }
